[IMP] sale_expense: Split file per models
This moves the account.move model into its own file, for more clarity

task-3458826
