+++ addons/sale_expense/models/__init__.py
@@ -1,5 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import account_move
 from . import account_move_line
 from . import hr_expense
 from . import hr_expense_sheet
